Tests/LibWeb/Text/input/css-has-invalidation/baseline-non-subject-has.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script src="_helpers.js"></script>
<style>
    .anchor:has(.match) .styled { color: red; }
</style>
<div id="a1">
    <div id="a2" class="anchor">
        <div id="a3">
            <span id="styled" class="styled">styled</span>
            <span id="target">target</span>
        </div>
    </div>
</div>
<script>
    test(() => {
        getComputedStyle(document.getElementById("styled")).color;
        internals.resetStyleInvalidationCounters();

        // Non-subject :has(): the :has() qualifies an ancestor, styling descendants.
        // Adding .match on target should invalidate the anchor path and re-style .styled.
        document.getElementById("target").classList.add("match");
        getComputedStyle(document.getElementById("styled")).color;
        printCounters("after add .match on target");
    });
</script>
